Separated Api from Main. Created hello template.
